Upgrade CI to macOS 26 and latest compilers (#169)
* Updates build workflow to use latest compilers

Configures the CI build job to use GCC 14 on Linux and Xcode 26.4 on macOS. This aligns the build environment with the benchmark workflow, ensuring that the main build process leverages modern toolchains for testing and optimization.

* Pins macOS runner to version 26 across CI workflows

Updates the build, multi-architecture, and language wrapper workflows to use the `macos-26` runner instead of `macos-latest`. This aligns the general CI environment with the benchmark workflow and ensures a consistent toolchain across all automated tasks.

* Updates language wrapper workflows to use latest compilers

Configures the Go, Node.js, Python, and Rust wrapper workflows to use GCC 14 on Linux and Xcode 26.4 on macOS. This ensures that language-specific builds and tests leverage the same modern toolchains as the core library and benchmark tasks.

* Updates macOS Intel runner to version 26

Updates the multi-architecture and language wrapper workflows to use the macos-26-intel runner. This aligns the Intel-based macOS CI jobs with the ARM64 jobs, ensuring a consistent build environment and toolchain across all macOS architectures.

* Fixes integer conversion warning in stream engine

Explicitly casts the size_t loop index to an integer when assigning the job ID. This silences compiler warnings regarding implicit type conversion and potential data loss on platforms where size_t is larger than int.

+++ src/lib/zxc_driver.c
@@ -596,7 +596,7 @@ static int64_t zxc_stream_engine_run(FILE* f_in, FILE* f_out, const int n_thread
     ZXC_MEMSET(buf_in, 0, ctx.ring_size * alloc_in);
 
     for (size_t i = 0; i < ctx.ring_size; i++) {
-        ctx.jobs[i].job_id = i;
+        ctx.jobs[i].job_id = (int)i;
         ctx.jobs[i].status = JOB_STATUS_FREE;
         ctx.jobs[i].in_buf = buf_in + (i * alloc_in);
         ctx.jobs[i].in_cap = alloc_in - ZXC_PAD_SIZE;
